Add create income with new category functionality

+++ Models/Expense.cs
@@ -6,6 +6,6 @@ public class ExpenseModel
 {
     public Guid expenseID {get;set;}
     public string description {get;set;}
-    public int amount {get;set;}
+    public float amount {get;set;}
     public string CategoryID {get;set;}
 }
