+++ Models/Income.cs
@@ -6,6 +6,6 @@ public class IncomeModel
 {
     public Guid incomeID {get;set;}
     public string description {get;set;}
-    public int amount {get;set;}
+    public float amount {get;set;}
     public string CategoryID {get;set;}
 }
